19w38a

+++ minecraft/src/net/minecraft/client/render/entity/model/EndermiteEntityModel.java
@@ -18,7 +18,10 @@ public class EndermiteEntityModel<T extends Entity> extends EntityModel<T> {
 
 		for (int i = 0; i < this.field_3368.length; i++) {
 			this.field_3368[i] = new ModelPart(this, field_3369[i][0], field_3369[i][1]);
-			this.field_3368[i].addCuboid((float)field_3366[i][0] * -0.5F, 0.0F, (float)field_3366[i][2] * -0.5F, field_3366[i][0], field_3366[i][1], field_3366[i][2]);
+			this.field_3368[i]
+				.addCuboid(
+					(float)field_3366[i][0] * -0.5F, 0.0F, (float)field_3366[i][2] * -0.5F, (float)field_3366[i][0], (float)field_3366[i][1], (float)field_3366[i][2]
+				);
 			this.field_3368[i].setRotationPoint(0.0F, (float)(24 - field_3366[i][1]), f);
 			if (i < this.field_3368.length - 1) {
 				f += (float)(field_3366[i][2] + field_3366[i + 1][2]) * 0.5F;
